pyproject.toml
[tool.poetry]
name = "sw-engineering-challenge"
version = "0.1.0"
description = "Package management for Bloq.it challenge"
authors = ["Breno Carvalho"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.4"
django-admin = "^2.0.2"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
